feat: Add more option

+++ src/main_funcs/handle_directdl.rs
@@ -49,10 +49,13 @@ pub async fn handle_direct(
     let ext = filepath.extension().and_then(|x| x.to_str()).unwrap();
     let out_name = format!(
         "{idxstr}{title}_[{id}].{ext}",
-        idxstr = if let Some(i) = i {
-            format!("{i:05}_")
+        idxstr = if !args.no_index_filename {
+            match i {
+                Some(index) => format!("{:05}_", index),
+                None => String::new(),
+            }
         } else {
-            "".to_string()
+            String::new()
         }
     );
 
@@ -74,7 +77,9 @@ pub async fn handle_direct(
 
     if let Some(op) = &op {
         copy_path_to_b2(&output_path, op).await?;
-        std::fs::remove_file(&output_path)?;
+        if !args.skip_video_delete {
+            std::fs::remove_file(&output_path)?;
+        }
     };
 
     Ok(())
